+++ src/pages/Admin/Community/CommunityUpdatesPage.tsx
@@ -2,10 +2,7 @@ import { Container } from '@mui/material';
 import React, { FC, useMemo } from 'react';
 import { useRouteMatch } from 'react-router-dom';
 import { WithCommunity } from '../../../components/Admin/Community/CommunityTypes';
-import {
-  CommunityUpdatesContainer,
-  CommunityUpdatesDataContainer,
-} from '../../../containers/community-updates/CommunityUpdates';
+import { CommunityUpdatesContainer } from '../../../containers/community-updates/CommunityUpdatesContainer';
 import { AvatarsProvider } from '../../../context/AvatarsProvider';
 import { useEcoverse, useUpdateNavigation } from '../../../hooks';
 import { CommunityUpdatesView } from '../../../views/CommunityUpdates/CommunityUpdatesView';
@@ -26,40 +23,31 @@ export const CommunityUpdatesPage: FC<CommunityUpdatesPageProps> = ({ paths, com
 
   return (
     <Container maxWidth="xl">
-      <CommunityUpdatesDataContainer
-        entities={{
-          ecoverseId,
-          communityId,
-        }}
-      >
-        {({ messages, senders }, { retrievingUpdateMessages }) => (
-          <CommunityUpdatesContainer entities={{ ecoverseId, communityId }}>
-            {(entities, actions, loading) => (
-              <AvatarsProvider users={senders}>
-                {populatedUsers => (
-                  <CommunityUpdatesView
-                    entities={{ messages, members: populatedUsers }}
-                    actions={{
-                      onSubmit: message => actions.onSubmit(message, communityId),
-                      onRemove: messageId => actions.onRemove(messageId, communityId),
-                    }}
-                    state={{
-                      loadingMessages: loading.retrievingUpdateMessages || retrievingUpdateMessages,
-                      submittingMessage: loading.sendingUpdateMessage,
-                      removingMessage: loading.removingUpdateMessage,
-                    }}
-                    options={{
-                      canEdit: true,
-                      canCopy: true,
-                      canRemove: true,
-                    }}
-                  />
-                )}
-              </AvatarsProvider>
+      <CommunityUpdatesContainer entities={{ ecoverseId, communityId }}>
+        {({ messages, senders }, actions, loading) => (
+          <AvatarsProvider users={senders}>
+            {populatedUsers => (
+              <CommunityUpdatesView
+                entities={{ messages, members: populatedUsers }}
+                actions={{
+                  onSubmit: message => actions.onSubmit(message, communityId),
+                  onRemove: messageId => actions.onRemove(messageId, communityId),
+                }}
+                state={{
+                  loadingMessages: loading.retrievingUpdateMessages,
+                  submittingMessage: loading.sendingUpdateMessage,
+                  removingMessage: loading.removingUpdateMessage,
+                }}
+                options={{
+                  canEdit: true,
+                  canCopy: true,
+                  canRemove: true,
+                }}
+              />
             )}
-          </CommunityUpdatesContainer>
+          </AvatarsProvider>
         )}
-      </CommunityUpdatesDataContainer>
+      </CommunityUpdatesContainer>
     </Container>
   );
 };
